requirements.txt
Pillow
discord.py
requests
opencv-python
numpy
ffmpeg-python
pydub
flask